backend/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cloudinary==1.44.1
cryptography==46.0.5
dj-database-url==3.1.2
Django==6.0.3
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
idna==3.11
MarkupSafe==3.0.3
pillow==12.1.1
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.12.1
python-dotenv==1.2.2
python-http-client==3.3.7
requests==2.32.5
sendgrid==6.12.5
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.3
Werkzeug==3.1.6
whitenoise==6.12.0
